+++ R/location.R
@@ -16,6 +16,8 @@
 #' changes(mtcars, mtcars)
 #' changes(mtcars, mtcars2)
 location <- function(df) {
+  signal_soft_deprecated("location() is deprecated, please use functions in the lobstr package")
+
   assert_that(is.data.frame(df))
 
   structure(list(
@@ -46,6 +48,8 @@ print.location <- function(x, ...) {
 #' @rdname location
 #' @export
 changes <- function(x, y) {
+  signal_soft_deprecated("changes() is deprecated, please use functions in the lobstr package")
+
   x <- location(x)
   y <- location(y)